+++ src/components/CommentInfo/CommentInfo.scss
