+++ src/components/CommentList/index.tsx
@@ -1,21 +1,1 @@
-import { CommentInfo } from '../CommentInfo';
-import { Comment } from '../../types/Comment';
-import './style.scss';
-
-type Props = {
-  comments: Comment[];
-};
-
-export const CommentList: React.FC<Props> = ({ comments }) => (
-  <ul data-cy="comments-list" className="CommentList">
-    {comments.map((comment) => (
-      <li key={comment.id}>
-        <CommentInfo
-          name={comment.name}
-          body={comment.body}
-          email={comment.email}
-        />
-      </li>
-    ))}
-  </ul>
-);
+export * from './CommentList';
