+++ src/components/PostInfo/PostInfo.scss
