+++ src/components/UserInfo/UserInfo.scss
